DCS-12: Remove AuthConfig to separate file.
