--- src/DiplomContentSystem/AuthConfig.cs
@@ -1,182 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.IdentityModel.Tokens;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace DiplomContentSystem
-{
-    public static class AuthConfig
-    {
-        public static IServiceCollection AddAuthPolicy(this IServiceCollection service)
-        {
-            return service.AddAuthorization(options =>
-            {
-                options.AddPolicy(AuthConsts.PolicyAdmin,
-                                  policy => policy.RequireClaim(AuthConsts.ClaimUserType, AuthConsts.RoleAdmin));
-                options.AddPolicy(AuthConsts.PolicyUser,
-                                  policy => policy.RequireClaim(AuthConsts.ClaimUserType, AuthConsts.RoleAdmin, AuthConsts.RoleUser));
-            });
-        }
-
-        public static IServiceCollection ConfigureJwtIssuerOptions(this IServiceCollection service, IConfiguration configuration, SymmetricSecurityKey _signingKey)
-        {
-            var jwtAppSettingOptions = configuration.GetSection(nameof(JwtIssuerOptions));
-
-            return service.Configure<JwtIssuerOptions>(options =>
-            {
-                options.Issuer = "DCS";//jwtAppSettingOptions[nameof(JwtIssuerOptions.Issuer)];
-                options.Audience = "DCS";// jwtAppSettingOptions[nameof(JwtIssuerOptions.Audience)];
-                options.SigningCredentials = new SigningCredentials(_signingKey, SecurityAlgorithms.HmacSha256);
-            });
-        }
-
-        public static IApplicationBuilder ConfigureJwtBearerAuthentication(this IApplicationBuilder app, IConfiguration configuration, SymmetricSecurityKey _signingKey)
-        {
-            var jwtAppSettingOptions = configuration.GetSection(nameof(JwtIssuerOptions));
-            var tokenValidationParameters = new TokenValidationParameters
-            {
-                ValidateIssuer = true,
-                ValidIssuer = "DCS",// jwtAppSettingOptions[nameof(JwtIssuerOptions.Issuer)],
-
-                ValidateAudience = true,
-                ValidAudience = "DCS",// jwtAppSettingOptions[nameof(JwtIssuerOptions.Audience)],
-
-                ValidateIssuerSigningKey = true,
-                IssuerSigningKey = _signingKey,
-
-                RequireExpirationTime = true,
-                ValidateLifetime = true,
-
-                ClockSkew = TimeSpan.Zero
-            };
-
-            return app.UseJwtBearerAuthentication(new JwtBearerOptions
-            {
-                AutomaticAuthenticate = true,
-                AutomaticChallenge = true,
-
-                TokenValidationParameters = tokenValidationParameters
-            });
-        }
-    }
-
-    internal class AuthConsts
-    {
-        public const string ClaimUserType = "UserType";
-
-        public const string PolicyAdmin = "AdminPolicy";
-
-        public const string PolicyUser = "UserPolicy";
-
-        public const string RoleUser = "User";
-
-        public const string RoleAdmin = "Admin";
-    }
-    public class JwtIssuerOptions
-    {
-        /// <summary>
-        /// "iss" (Issuer) Claim
-        /// </summary>
-        /// <remarks>The "iss" (issuer) claim identifies the principal that issued the
-        ///   JWT.  The processing of this claim is generally application specific.
-        ///   The "iss" value is a case-sensitive string containing a StringOrURI
-        ///   value.  Use of this claim is OPTIONAL.</remarks>
-        public string Issuer { get; set; }
-
-        /// <summary>
-        /// "sub" (Subject) Claim
-        /// </summary>
-        /// <remarks> The "sub" (subject) claim identifies the principal that is the
-        ///   subject of the JWT.  The claims in a JWT are normally statements
-        ///   about the subject.  The subject value MUST either be scoped to be
-        ///   locally unique in the context of the issuer or be globally unique.
-        ///   The processing of this claim is generally application specific.  The
-        ///   "sub" value is a case-sensitive string containing a StringOrURI
-        ///   value.  Use of this claim is OPTIONAL.</remarks>
-        public string Subject { get; set; }
-
-        /// <summary>
-        /// "aud" (Audience) Claim
-        /// </summary>
-        /// <remarks>The "aud" (audience) claim identifies the recipients that the JWT is
-        ///   intended for.  Each principal intended to process the JWT MUST
-        ///   identify itself with a value in the audience claim.  If the principal
-        ///   processing the claim does not identify itself with a value in the
-        ///   "aud" claim when this claim is present, then the JWT MUST be
-        ///   rejected.  In the general case, the "aud" value is an array of case-
-        ///   sensitive strings, each containing a StringOrURI value.  In the
-        ///   special case when the JWT has one audience, the "aud" value MAY be a
-        ///   single case-sensitive string containing a StringOrURI value.  The
-        ///   interpretation of audience values is generally application specific.
-        ///   Use of this claim is OPTIONAL.</remarks>
-        public string Audience { get; set; }
-
-        /// <summary>
-        /// "nbf" (Not Before) Claim (default is UTC NOW)
-        /// </summary>
-        /// <remarks>The "nbf" (not before) claim identifies the time before which the JWT
-        ///   MUST NOT be accepted for processing.  The processing of the "nbf"
-        ///   claim requires that the current date/time MUST be after or equal to
-        ///   the not-before date/time listed in the "nbf" claim.  Implementers MAY
-        ///   provide for some small leeway, usually no more than a few minutes, to
-        ///   account for clock skew.  Its value MUST be a number containing a
-        ///   NumericDate value.  Use of this claim is OPTIONAL.</remarks>
-        public DateTime NotBefore => DateTime.UtcNow;
-
-        /// <summary>
-        /// "iat" (Issued At) Claim (default is UTC NOW)
-        /// </summary>
-        /// <remarks>The "iat" (issued at) claim identifies the time at which the JWT was
-        ///   issued.  This claim can be used to determine the age of the JWT.  Its
-        ///   value MUST be a number containing a NumericDate value.  Use of this
-        ///   claim is OPTIONAL.</remarks>
-        public DateTime IssuedAt => DateTime.UtcNow;
-
-        /// <summary>
-        /// Set the remember timespan the token will be valid for (default is 10 days)
-        /// </summary>
-        public TimeSpan RememberValidFor { get; set; } = TimeSpan.FromDays(10);
-
-        /// <summary>
-        /// Set the timespan the token will be valid for (default is 30 min)
-        /// </summary>
-        public TimeSpan ValidFor { get; set; } = TimeSpan.FromMinutes(30);
-
-        /// <summary>
-        /// "exp" (Expiration Time) Claim (returns IssuedAt + ValidFor)
-        /// </summary>
-        /// <remarks>The "exp" (expiration time) claim identifies the expiration time on
-        ///   or after which the JWT MUST NOT be accepted for processing.  The
-        ///   processing of the "exp" claim requires that the current date/time
-        ///   MUST be before the expiration date/time listed in the "exp" claim.
-        ///   Implementers MAY provide for some small leeway, usually no more than
-        ///   a few minutes, to account for clock skew.  Its value MUST be a number
-        ///   containing a NumericDate value.  Use of this claim is OPTIONAL.</remarks>
-        public DateTime RememberExpiration => IssuedAt.Add(RememberValidFor);
-
-        public DateTime Expiration => IssuedAt.Add(ValidFor);
-
-        /// <summary>
-        /// "jti" (JWT ID) Claim (default ID is a GUID)
-        /// </summary>
-        /// <remarks>The "jti" (JWT ID) claim provides a unique identifier for the JWT.
-        ///   The identifier value MUST be assigned in a manner that ensures that
-        ///   there is a negligible probability that the same value will be
-        ///   accidentally assigned to a different data object; if the application
-        ///   uses multiple issuers, collisions MUST be prevented among values
-        ///   produced by different issuers as well.  The "jti" claim can be used
-        ///   to prevent the JWT from being replayed.  The "jti" value is a case-
-        ///   sensitive string.  Use of this claim is OPTIONAL.</remarks>
-        public Func<Task<string>> JtiGenerator =>
-          () => Task.FromResult(Guid.NewGuid().ToString());
-
-        /// <summary>
-        /// The signing key to use when generating tokens.
-        /// </summary>
-        public SigningCredentials SigningCredentials { get; set; }
-    }
-}
